Prevent calendar error when inserting forms with Live Composer
Live Composer will load the HTML for the form, insert it into the page,
and then save all changes.

However, when HTML is dynamically added to the page, there is no longer
a guarantee that scripts will run in order. This causes errors in live
composer, which means it will not save the changes.

We should load in the calendar JS ahead of time, so that there will not
be any script errors.

This is a bit messy, but I don’t think there is another real solution
available.

[#114529991]