Fixes for SSL issue:
Instead of caching the form HTML source code in the WordPress database, we now generate it on every page load, and determine the protocol and domain at the same time, based on how the page is being accessed (HTTP or HTTPS).

Also on the AC settings page (in WordPress admin) I added another form submit button near the top (it's hard to see at the bottom only when you have a lot of forms). Also updated the screenshot image when the AC settings are completely disconnected.

[#115362467]